prediction-ia-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2
torch==2.1.1
torch-geometric==2.4.0
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6


